refactor(claim): extract ClaimQuizSheet for reuse by route mode
Pulls the searching→results→empty/quiz/quizFailed/claimed state machine
out of Claim into a new reusable ClaimQuizSheet widget
(lib/widgets/claim_quiz_sheet.dart). Claim now owns only the initial
state (scan button + streak badge) and delegates everything from the
nearbyPostboxes call onward to the sheet via onCompleted/onCancel
callbacks. ClaimQuizResult carries claimedCount/pointsEarned/quizFailed/
empty back to the parent. The compact flag scales up touch targets for
the upcoming LiveRouteScreen bottom-sheet variant (Task 8).

Behaviour parity preserved: same Cloud Function calls, quiz option
generation, James messages, analytics events, confetti, and animations.
